agent/requirements.txt
# Digital Archivist Agent - Python 依赖

# Web3 交互
web3==6.11.1

# OpenAI API
openai==1.3.0

# 环境变量管理
python-dotenv==1.0.0

# 可选：使用 Anthropic Claude 替代 OpenAI
# anthropic==0.5.0

# JSON 处理（Python 内置，无需安装）
# json

# 可选：IPFS 集成
# ipfshttpclient==0.8.0a2


